requirements.txt
byaldi
pillow
PyMuPDF
torch